+++ apps/meteor/app/custom-oauth/server/custom_oauth_server.d.ts
@@ -2,4 +2,6 @@ export class CustomOAuth {
 	constructor(name: string, options: Record<string, any>);
 
 	getIdentity(accessToken: string, query: Record<string, any>): any;
+
+	configure(options: Record<string, any>): any;
 }
