+++ apps/meteor/app/dolphin/client/index.ts
@@ -1,2 +1,2 @@
-import '../lib/common';
+import './lib';
 import './login-button.css';
